# openapi.yaml
openapi: 3.0.3
info:
  title: Building Server API
  description: RESTful API for user management with authentication and authorization
  version: 0.1.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: https://localhost:8443
    description: Production HTTPS server
  - url: http://localhost:8080
    description: Development HTTP server
...
components:
  schemas:
    Response:
      type: object
      required:
        - msg
      properties:
        msg:
          type: string
          description: Response message
        data:
          description: Optional response data
    ErrorResponse:
      type: object
      required:
        - msg
      properties:
        msg:
          type: string
          description: Error message
    RegisterRequest:
      type: object
      required:
        - email
        - username
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        username:
          type: string
          minLength: 3
          maxLength: 30
          description: Username
          example: johndoe
        password:
          type: string
          minLength: 8
          format: password
          description: User password (at least 8 characters)
          example: securePassword123
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        password:
          type: string
          format: password
          description: User password
          example: securePassword123
    TokenResponse:
      allOf:
        - $ref: '#/components/schemas/Response'
        - type: object
          properties:
            data:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  description: JWT authentication token
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
...
    UserInfo:
      type: object
      required:
        - id
        - email
        - username
        - is_admin
      properties:
        id:
          type: string
          description: User ObjectId
          example: 507f1f77bcf86cd799439011
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        username:
          type: string
          description: Username
          example: johndoe
        is_admin:
          type: boolean
          description: Admin status
          example: false
    UserInfoResponse:
      allOf:
        - $ref: '#/components/schemas/Response'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UserInfo'
    UserListResponse:
      allOf:
        - $ref: '#/components/schemas/Response'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/UserInfo'
    CreateUserRequest:
      type: object
      required:
        - email
        - username
        - password
        - is_admin
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: newuser@example.com
        username:
          type: string
          minLength: 3
          maxLength: 30
          description: Username
          example: newuser
        password:
          type: string
          minLength: 8
          format: password
          description: User password (at least 8 characters)
          example: password123
        is_admin:
          type: boolean
          description: Admin status
          example: false
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: New email address (optional)
          example: updated@example.com
        username:
          type: string
          minLength: 3
          maxLength: 30
          description: New username (optional)
          example: updateduser
        password:
          type: string
          minLength: 8
          format: password
          description: New password (optional, at least 8 characters)
          example: newPassword123
    SetRoleRequest:
      type: object
      required:
        - is_admin
      properties:
        is_admin:
          type: boolean
          description: Set admin status
          example: true
...
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            msg: Invalid or expired token
    Forbidden:
      description: Forbidden - Admin privileges required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            msg: Admin access required
    BadRequest:
      description: Bad Request - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            msg: Validation error
    NotFound:
      description: Not Found - Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            msg: User not found
    Conflict:
      description: Conflict - Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            msg: Email already exists
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and receive JWT token
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Invalidate current JWT token (adds to blacklist)
      operationId: logout
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                msg: Logout successful
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /admin/users:
    get:
      tags:
        - Admin
      summary: Get all users
      description: Retrieve a list of all users (admin only)
      operationId: getAllUsers
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Admin
      summary: Create new user
      description: Create a new user account (admin only)
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                msg: User created successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/users/{id}:
    get:
      tags:
        - Admin
      summary: Get user by ID
      description: Retrieve a specific user's information by ID (admin only)
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          description: User ObjectId
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Admin
      summary: Update user
      description: Update a user's information (admin only)
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          description: User ObjectId
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                msg: User updated successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

    delete:
      tags:
        - Admin
      summary: Delete user
      description: Delete a user account (admin only)
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          description: User ObjectId
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                msg: User deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/{id}/admin:
    put:
      tags:
        - Admin
      summary: Set user admin role
      description: Grant or revoke admin privileges for a user (admin only)
      operationId: setAdmin
      parameters:
        - name: id
          in: path
          required: true
          description: User ObjectId
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetRoleRequest'
      responses:
        '200':
          description: User role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                msg: User set as admin
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/books:
    post:
      tags:
        - Admin
      summary: Create new book
      description: Create a new book (admin only)
      operationId: createBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookRequest'
      responses:
        '201':
          description: Book created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                msg: successfully created book
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /admin/books/{id}:
    put:
      tags:
        - Admin
      summary: Update book
      description: Update a book's information (admin only)
      operationId: updateBook
      parameters:
        - name: id
          in: path
          required: true
          description: Book ObjectId
          schema:
            type: string
            example: 507f1f77bcf86cd799439012
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBookRequest'
      responses:
        '200':
          description: Book updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                msg: successfully updated book
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Admin
      summary: Delete book
      description: Delete a book (admin only)
      operationId: deleteBook
      parameters:
        - name: id
          in: path
          required: true
          description: Book ObjectId
          schema:
            type: string
            example: 507f1f77bcf86cd799439012
      responses:
        '200':
          description: Book deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                msg: successfully deleted book
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...